+++ internal/entities/order-comments-entity.go
@@ -2,14 +2,12 @@ package entities
 
 import "request-system/pkg/types"
 
-
 type OrderComment struct {
-	ID           int    `json:"id"`
-	Message      string `json:"message"`
-	StatusID     int    `json:"status_id"`
-	OrderID      int    `json:"order_id"`
-	UserID       int    `json:"user_id"`
+	ID       uint64 `json:"id"`
+	Message  string `json:"message"`
+	StatusID uint64 `json:"status_id"`
+	OrderID  uint64 `json:"order_id"`
+	UserID   uint64 `json:"user_id"`
 
-	types.BaseEntity 
-	
+	types.BaseEntity
 }
